Taste.com for Educators.. Google Maps for Curriculum..
I'm looking for direction.. to create visual curriculum (for educators) based on Australian Curriculum. I understand what I want is called 'data visualisation'. (OK, I'm really new to this). It needs to be visual, interactive, fast, beautifully simple - and compelling! It's what teachers needs to help them make sense of the vast amount of 'stuff' they need to manage, in the course of their curriculum planning and teaching. I'm looking for advice; people to share ideas with; get advice - and, hopefully, someone to help me develop this project.